Added utils to check id and name existence and implemented into routes

+++ backend/routes/nonprofit.js
@@ -1,76 +1,83 @@
 import { PrismaClient } from "#prisma/client.js";
 import express from "express";
+import {
+  checkNonProfitName,
+  checkNonProfitId,
+} from "#utils/nonprofit_utils.js";
 const prisma = new PrismaClient();
 const nonprofitRouter = express.Router();
 
 // Default is to get all nonprofits
 nonprofitRouter.get("/", async (req, res) => {
-  const findNonProfits = await prisma.nonprofit.findMany();
-  res.status(200).json(findNonProfits);
+  const foundNonProfits = await prisma.nonprofit.findMany();
+  res.status(200).json(foundNonProfits);
 });
 
 // Get one nonprofit by name
-nonprofitRouter.get("/:nonprofitname", async (req, res) => {
-  const { nonprofitname } = req.params;
+nonprofitRouter.get("/:nonprofit_id", async (req, res) => {
+  const { nonprofit_id } = req.params;
   const findNonProfit = await prisma.nonprofit.findUnique({
     where: {
-      name: nonprofitname,
+      id: nonprofit_id,
     },
   });
-  res.status(200).json(findNonProfit);
+  if (findNonProfit) {
+    res.status(200).json(findNonProfit);
+  } else {
+    res.status(404).send(`Nonprofit ${nonprofit_id} not found`);
+  }
 });
 
 // Add a new nonprofit by name
 nonprofitRouter.post("/add", async (req, res) => {
   const nonProfitData = req.body;
   const name = nonProfitData.name;
+  const exists = await checkNonProfitName(name);
 
-  const findNonProfits = await prisma.nonprofit.findUnique({
-    where: {
-      name: name,
-    },
-  });
-  if (!findNonProfits) {
+  if (!exists) {
     const createNonProfit = await prisma.nonprofit.create({
       data: nonProfitData,
     });
     res.status(201).json(createNonProfit);
   } else {
-    req.status(400).json({ message: "Nonprofit already exists" });
+    res.status(400).send(`Nonprofit ${name} already exists`);
   }
 });
 
 // Edit a nonprofit by name
-nonprofitRouter.put("/:nonprofitname/edit", async (req, res) => {
-  const { nonprofitname } = req.params;
+nonprofitRouter.put("/:nonprofit_id/edit", async (req, res) => {
+  const { nonprofit_id } = req.params;
   const nonProfitData = req.body;
-  try {
+  const exists = await checkNonProfitId(nonprofit_id);
+
+  if (exists) {
     const updateOne = await prisma.nonprofit.update({
       where: {
-        name: nonprofitname,
+        id: nonprofit_id,
       },
-
       data: nonProfitData,
     });
     res.json(updateOne);
     res.status(214).send();
-  } catch (PrismaClientKnownRequestError) {
-    res.status(404).send("Nonprofit not found");
+  } else {
+    res.status(404).send(`Nonprofit ${nonprofit_id} not found`);
   }
 });
 
 // Delete a nonprofit by name
-nonprofitRouter.delete("/:nonprofitname/delete", async (req, res) => {
-  const { nonprofitname } = req.params;
-  try {
+nonprofitRouter.delete("/:nonprofit_id/delete", async (req, res) => {
+  const { nonprofit_id } = req.params;
+  const exists = await checkNonProfitId(nonprofit_id);
+
+  if (exists) {
     await prisma.nonprofit.delete({
       where: {
-        name: nonprofitname,
+        id: nonprofit_id,
       },
     });
-    res.status(214).send();
-  } catch (PrismaClientKnownRequestError) {
-    res.status(404).send("Nonprofit not found");
+    res.status(214).send(`Nonprofit ${nonprofit_id} deleted`);
+  } else {
+    res.status(404).send(`Nonprofit ${nonprofit_id} not found`);
   }
 });
